clean up repos methods, move duplicate into the centralize localtion
