+++ nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/rdb/dao/RdbDataPersistentDAO.java
@@ -61,421 +61,25 @@ public class RdbDataPersistentDAO {
         this.jdbcTemplateUtil = jdbcTemplateUtil;
     }
 
-
-    @SuppressWarnings("java:S3776")
-    protected void persistingGenericTable(String jsonData, boolean initialLoad,
-                                          PollDataSyncConfig config) {
-        if (config.getTableName() != null && !config.getTableName().isEmpty()) {
-            SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate);
-            simpleJdbcInsert = simpleJdbcInsert.withTableName(config.getTableName());
-            List<Map<String, Object>> records = PollServiceUtil.jsonToListOfMap(jsonData);
-
-            if (records != null && !records.isEmpty()) {
-                // Handle special filtering for specified tables
-                Set<String> specialTables = new HashSet<>(Arrays.asList(
-                        "D_INV_HIV",
-                        "D_INV_ADMINISTRATIVE",
-                        "D_INV_EPIDEMIOLOGY",
-                        "D_INV_LAB_FINDING",
-                        "D_INV_MEDICAL_HISTORY",
-                        "D_INV_RISK_FACTOR",
-                        "D_INV_TREATMENT",
-                        "D_INV_VACCINATION"
-                ));
-
-                if (specialTables.contains(config.getTableName().toUpperCase()))
-                {
-                    handleSpecialTableFiltering(config, records, simpleJdbcInsert, initialLoad);
-                }
-                else
-                {
-                    // For other tables, use the existing logic
-                    try {
-                        if (records.size() > batchSize) {
-                            int sublistSize = batchSize;
-                            for (int i = 0; i < records.size(); i += sublistSize) {
-                                int end = Math.min(i + sublistSize, records.size());
-                                List<Map<String, Object>> sublist = records.subList(i, end);
-                                if (initialLoad || config.getKeyList().isEmpty()) {
-                                    sublist.forEach(data -> data.remove("RowNum"));
-                                    simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(sublist));
-
-                                } else {
-                                    jdbcTemplateUtil.upsertBatch(config.getTableName(), sublist, config.getKeyList());
-                                }
-                            }
-                        } else {
-//                            simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(records));
-                            if (initialLoad || config.getKeyList().isEmpty()) {
-                                records.forEach(data -> data.remove("RowNum"));
-                                simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(records));
-
-                            } else {
-                                jdbcTemplateUtil.upsertBatch(config.getTableName(), records, config.getKeyList());
-                            }
-                        }
-                    } catch (Exception e) {
-                        handleBatchInsertionFailure(records, config.getTableName(), simpleJdbcInsert);
-                    }
-                }
-            }
-        }
-    }
-
-    // Helper method to handle filtering for special tables
-    @SuppressWarnings("java:S3776")
-    protected void handleSpecialTableFiltering(PollDataSyncConfig config, List<Map<String, Object>> records, SimpleJdbcInsert simpleJdbcInsert, boolean initialLoad) {
-        String keyColumn = config.getTableName() + "_KEY"; // Assuming each table has a key column with the pattern [table]_KEY
-
-        // Query only the existing keys where the key is 1
-        String query = "SELECT " + keyColumn + " FROM " + config.getTableName() + " WHERE " + keyColumn + " = 1";
-        List<Double> existingKeys = new ArrayList<>(jdbcTemplate.queryForList(query, Double.class));
-        logger.info("Found {} existing keys in {} with {} = 1", existingKeys.size(), config.getTableName(), keyColumn);
-
-        Double foundKey = existingKeys.isEmpty()? null : existingKeys.get(0);
-        // Filter out records that have a matching key of 1
-        List<Map<String, Object>> filteredRecords = records
-                .stream()
-                .filter(rec -> {
-                    // Check if the record contains the keyColumn
-                    if (rec.containsKey(keyColumn)) {
-                        // Get the value of the keyColumn from the record
-                        Object recordKey = rec.get(keyColumn);
-                        // Filter out the record if it matches the foundKey
-                        return !recordKey.equals(foundKey);
-                    }
-                    // If the record doesn't contain keyColumn, keep it
-                    return true;
-                })
-                .collect(Collectors.toList()); //NOSONAR
-        logger.info("After filtering, Records size: {}", filteredRecords.size());
-
-        // Additional logic block to remove records with duplicate keys based on keyColumn
-        Set<Object> uniqueKeys = new HashSet<>();
-        List<Map<String, Object>> deduplicatedRecords = new ArrayList<>();
-
-        for (Map<String, Object> rec : filteredRecords) {
-            Object recordKey = rec.get(keyColumn);
-
-            // If the keyColumn is present and the key hasn't been seen yet, keep the record
-            if (recordKey != null && uniqueKeys.add(recordKey)) {
-                deduplicatedRecords.add(rec);
-            }
-        }
-
-
-        if (!deduplicatedRecords.isEmpty()) {
-            try {
-                if (deduplicatedRecords.size() > batchSize) {
-                    int sublistSize = batchSize;
-                    for (int i = 0; i < deduplicatedRecords.size(); i += sublistSize) {
-                        int end = Math.min(i + sublistSize, deduplicatedRecords.size());
-                        List<Map<String, Object>> sublist = deduplicatedRecords.subList(i, end);
-//                        simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(sublist));
-                        if (initialLoad || config.getKeyList().isEmpty()) {
-                            sublist.forEach(data -> data.remove("RowNum"));
-                            simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(sublist));
-
-                        } else {
-                            jdbcTemplateUtil.upsertBatch(config.getTableName(), sublist, config.getKeyList());
-                        }
-                    }
-                } else {
-//                    simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(deduplicatedRecords));
-                    if (initialLoad || config.getKeyList().isEmpty()) {
-                        records.forEach(data -> data.remove("RowNum"));
-                        simpleJdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(records));
-
-                    } else {
-                        jdbcTemplateUtil.upsertBatch(config.getTableName(), records, config.getKeyList());
-                    }
-                }
-            } catch (Exception e) {
-                handleBatchInsertionFailure(records, config.getTableName(), simpleJdbcInsert);
-            }
-        } else {
-            logger.info("No new records to insert for {}.", config.getTableName());
-        }
-    }
-
-
-    // Helper method to handle batch insertion failure
-    protected void handleBatchInsertionFailure(List<Map<String, Object>> records, String tableName, SimpleJdbcInsert simpleJdbcInsert) {
-        for (Map<String, Object> res : records) {
-            try {
-                simpleJdbcInsert.execute(new MapSqlParameterSource(res));
-            } catch (Exception ei) {
-                logger.error("ERROR occurred at record: {}, {}", gsonNorm.toJson(res), ei.getMessage()); // NOSONAR
-                handleError.writeRecordToFile(gsonNorm, res, tableName + UUID.randomUUID(), sqlErrorPath + "/RDB/" + ei.getClass().getSimpleName() + "/" + tableName + "/");
-            }
-        }
-    }
-
     @SuppressWarnings({"java:S3776", "java:S125"})
     public LogResponseModel saveRDBData(PollDataSyncConfig config, String jsonData, boolean initialLoad) {
         LogResponseModel logBuilder = new LogResponseModel();
         StringBuilder builder = new StringBuilder();
-//        if ("CONFIRMATION_METHOD".equalsIgnoreCase(config.getTableName())) {
-//            logBuilder.setLog(LOG_SUCCESS);
-//            logBuilder.setStatus(SUCCESS);
-//            Type resultType = new TypeToken<List<ConfirmationMethod>>() {}.getType();
-//            List<ConfirmationMethod> list = gson.fromJson(jsonData, resultType);
-//
-//
-//            for (ConfirmationMethod confirmationMethod : list) {
-//                builder.append(", ").append(upsertConfirmationMethod(confirmationMethod));
-//            }
-//        } else if ("CONDITION".equalsIgnoreCase(config.getTableName())) {
-//            logBuilder.setLog(LOG_SUCCESS);
-//            logBuilder.setStatus(SUCCESS);
-//            Type resultType = new TypeToken<List<Condition>>() {
-//            }.getType();
-//            List<Condition> list = gson.fromJson(jsonData, resultType);
-//            for (Condition condition : list) {
-//                builder.append(", ").append(upsertCondition(condition));
-//            }
-//        } else if ("RDB_DATE".equalsIgnoreCase(config.getTableName())) {
-//            logBuilder.setLog(LOG_SUCCESS);
-//            logBuilder.setStatus(SUCCESS);
-//            Type resultType = new TypeToken<List<RdbDate>>() {
-//            }.getType();
-//            List<RdbDate> list = gson.fromJson(jsonData, resultType);
-//            for (RdbDate rdbDate : list) {
-//                builder.append(", ").append(upsertRdbDate(rdbDate));
-//            }
-//        } else {
-            try {
-                persistingGenericTable(jsonData, initialLoad, config);
-            }
-            catch (Exception e) {
-                logBuilder =  new LogResponseModel();
-                logBuilder.setLog(e.getMessage());
-                logBuilder.setStatus(ERROR);
-                logBuilder.setStackTrace(getStackTraceAsString(e));
-                logger.error("Error executeBatch. class: {}, tableName: {}, Error:{}", e.getClass() ,config.getTableName(), e.getMessage());
-            }
-//        }
+
+        try {
+            jdbcTemplateUtil.persistingGenericTable(jsonData,config, initialLoad);
+        }
+        catch (Exception e) {
+            logBuilder =  new LogResponseModel();
+            logBuilder.setLog(e.getMessage());
+            logBuilder.setStatus(ERROR);
+            logBuilder.setStackTrace(getStackTraceAsString(e));
+            logger.error("Error executeBatch. class: {}, tableName: {}, Error:{}", e.getClass() ,config.getTableName(), e.getMessage());
+        }
+
 
         logBuilder.setLog(builder.toString());
         return logBuilder;
     }
 
-    private String upsertConfirmationMethod(ConfirmationMethod confirmationMethod) {
-        String log = LOG_SUCCESS;
-        String sql = "MERGE INTO CONFIRMATION_METHOD AS target USING " +
-                "(select " + confirmationMethod.getConfirmationMethodKey() + " as id) AS source " +
-                "ON (target.CONFIRMATION_METHOD_KEY = source.id) " +
-                "WHEN MATCHED THEN " +
-                " UPDATE SET target.CONFIRMATION_METHOD_CD = " + getSqlString(confirmationMethod.getConfirmationMethodCd()) +
-                ",target.CONFIRMATION_METHOD_DESC = " + getSqlString(confirmationMethod.getConfirmationMethodDesc()) +
-                " WHEN NOT MATCHED THEN " +
-                " INSERT (CONFIRMATION_METHOD_KEY, CONFIRMATION_METHOD_CD,CONFIRMATION_METHOD_DESC) " +
-                "VALUES (" + confirmationMethod.getConfirmationMethodKey() +
-                "," + getSqlString(confirmationMethod.getConfirmationMethodCd()) +
-                "," + getSqlString(confirmationMethod.getConfirmationMethodDesc()) + ");";
-        try {
-            jdbcTemplate.update(sql);
-        } catch (Exception e) {
-            logger.error("Error in upsert for CONFIRMATION_METHOD table:{}", e.getMessage());
-            log = e.getMessage();
-        }
-        return log;
-    }
-
-    protected String upsertCondition(Condition condition) {
-        String sql = "MERGE INTO CONDITION AS target " +
-                "USING (select " + condition.getConditionKey() + " as id) AS source ON" +
-                "(target.CONDITION_KEY = source.id)" +
-                "WHEN MATCHED THEN " +
-                "UPDATE " +
-                "SET " +
-                " target.CONDITION_CD = " + getSqlString(condition.getConditionCd()) + "," +
-                " target.CONDITION_DESC = " + getSqlString(condition.getConditionDesc()) + "," +
-                " target.CONDITION_SHORT_NM = " + getSqlString(condition.getConditionShortNm()) + "," +
-                " target.CONDITION_CD_EFF_DT = " + getSqlString(condition.getConditionCdEffDt()) + "," +
-                " target.CONDITION_CD_END_DT = " + getSqlString(condition.getConditionCdEndDt()) + "," +
-                " target.NND_IND = " + getSqlString(condition.getNndInd()) + "," +
-                " target.DISEASE_GRP_CD = " + getSqlString(condition.getDiseaseGrpCd()) + "," +
-                " target.DISEASE_GRP_DESC = " + getSqlString(condition.getDiseaseGrpDesc()) + "," +
-                " target.PROGRAM_AREA_CD = " + getSqlString(condition.getProgramAreaCd()) + "," +
-                " target.PROGRAM_AREA_DESC = " + getSqlString(condition.getProgramAreaDesc()) + "," +
-                " target.CONDITION_CD_SYS_CD_NM = " + getSqlString(condition.getConditionCdSysCdNm()) + "," +
-                " target.ASSIGNING_AUTHORITY_CD = " + getSqlString(condition.getAssigningAuthorityCd()) + "," +
-                " target.ASSIGNING_AUTHORITY_DESC = " + getSqlString(condition.getAssigningAuthorityDesc()) + "," +
-                " target.CONDITION_CD_SYS_CD = " + getSqlString(condition.getConditionCdSysCd()) +
-                " WHEN NOT MATCHED THEN " +
-                "INSERT " +
-                " (CONDITION_KEY," +
-                " CONDITION_CD," +
-                " CONDITION_DESC,CONDITION_SHORT_NM,CONDITION_CD_EFF_DT,CONDITION_CD_END_DT,NND_IND," +
-                " DISEASE_GRP_CD,DISEASE_GRP_DESC,PROGRAM_AREA_CD,PROGRAM_AREA_DESC," +
-                " CONDITION_CD_SYS_CD_NM,ASSIGNING_AUTHORITY_CD,ASSIGNING_AUTHORITY_DESC,CONDITION_CD_SYS_CD)" +
-                "VALUES ("
-                + condition.getConditionKey() + "," +
-                getSqlString(condition.getConditionCd()) + "," +
-                getSqlString(condition.getConditionDesc()) + "," +
-                getSqlString(condition.getConditionShortNm()) + "," +
-                getSqlString(condition.getConditionCdEffDt()) + "," +
-                getSqlString(condition.getConditionCdEndDt()) + "," +
-                getSqlString(condition.getNndInd()) + "," +
-                getSqlString(condition.getDiseaseGrpCd()) + "," +
-                getSqlString(condition.getDiseaseGrpDesc()) + "," +
-                getSqlString(condition.getProgramAreaCd()) + "," +
-                getSqlString(condition.getProgramAreaDesc()) + "," +
-                getSqlString(condition.getConditionCdSysCdNm()) + "," +
-                getSqlString(condition.getAssigningAuthorityCd()) + "," +
-                getSqlString(condition.getAssigningAuthorityDesc()) + "," +
-                getSqlString(condition.getConditionCdSysCd()) +
-                ");";
-        String log = LOG_SUCCESS;
-        try {
-            jdbcTemplate.update(sql);
-        } catch (Exception e) {
-            logger.error("Error in upsert for CONDITION table:{}", e.getMessage());
-            log = e.getMessage();
-        }
-        return log;
-    }
-
-    protected String upsertRdbDate(RdbDate rdbDate) {
-        String sql = "MERGE INTO RDB_DATE AS target " +
-                "USING (select " + rdbDate.getDateKey() + " as id) AS source ON" +
-                "(target.DATE_KEY = source.id)" +
-                "WHEN MATCHED THEN " +
-                "UPDATE " +
-                "SET " +
-                " target.DATE_MM_DD_YYYY = " + getSqlString(rdbDate.getDateMmDdYyyy()) + "," +
-                " target.DAY_OF_WEEK = " + getSqlString(rdbDate.getDayOfWeek()) + "," +
-                " target.DAY_NBR_IN_CLNDR_MON = " + rdbDate.getDayNbrInClndrMon() + "," +
-                " target.DAY_NBR_IN_CLNDR_YR = " + rdbDate.getDayNbrInClndrYr() + "," +
-                " target.WK_NBR_IN_CLNDR_MON = " + rdbDate.getWkNbrInClndrMon() + "," +
-                " target.WK_NBR_IN_CLNDR_YR = " + rdbDate.getWkNbrInClndrYr() + "," +
-                " target.CLNDR_MON_NAME = " + getSqlString(rdbDate.getClndrMonName()) + "," +
-                " target.CLNDR_MON_IN_YR = " + rdbDate.getClndrMonInYr() + "," +
-                " target.CLNDR_QRTR = " + rdbDate.getClndrQrtr() + "," +
-                " target.CLNDR_YR = " + rdbDate.getClndrYr() +
-                " WHEN NOT MATCHED THEN " +
-                "INSERT " +
-                " (DATE_KEY," +
-                " DATE_MM_DD_YYYY," +
-                " DAY_OF_WEEK,DAY_NBR_IN_CLNDR_MON,DAY_NBR_IN_CLNDR_YR,WK_NBR_IN_CLNDR_MON,WK_NBR_IN_CLNDR_YR," +
-                " CLNDR_MON_NAME,CLNDR_MON_IN_YR,CLNDR_QRTR,CLNDR_YR)" +
-                " VALUES ("
-                + rdbDate.getDateKey() + "," +
-                getSqlString(rdbDate.getDateMmDdYyyy()) + "," +
-                getSqlString(rdbDate.getDayOfWeek()) + "," +
-                rdbDate.getDayNbrInClndrMon() + "," +
-                rdbDate.getDayNbrInClndrYr() + "," +
-                rdbDate.getWkNbrInClndrMon() + "," +
-                rdbDate.getWkNbrInClndrYr() + "," +
-                getSqlString(rdbDate.getClndrMonName()) + "," +
-                rdbDate.getClndrMonInYr() + "," +
-                rdbDate.getClndrQrtr() + "," +
-                rdbDate.getClndrYr() +
-                ");";
-
-        String log = LOG_SUCCESS;
-        try {
-            jdbcTemplate.update(sql);
-        } catch (Exception e) {
-            logger.error("Error in upsert for RDB_DATE table:{}", e.getMessage());
-            log = e.getMessage();
-        }
-        return log;
-    }
-
-    private static String getSqlString(String val) {
-        if (val != null && !val.isEmpty()) {
-            return "'" + val + "'";
-        }
-        return val;
-    }
-
-    public String getLastUpdatedTime(String tableName) {
-        String sql = "select last_update_time from POLL_DATA_SYNC_CONFIG where table_name=?";
-        String updatedTime = "";
-        Timestamp lastTime = this.jdbcTemplate.queryForObject(
-                sql,
-                Timestamp.class, tableName);
-        if (lastTime != null) {
-            SimpleDateFormat formatter = new SimpleDateFormat(TIMESTAMP_FORMAT);
-            updatedTime = formatter.format(lastTime); // NOSONAR
-        }
-        return updatedTime;
-    }
-
-    public String getLastUpdatedTimeS3(String tableName) {
-        String sql = "select last_update_time_s3 from POLL_DATA_SYNC_CONFIG where table_name=?";
-        String updatedTime = "";
-        Timestamp lastTime = this.jdbcTemplate.queryForObject(
-                sql,
-                Timestamp.class, tableName);
-        if (lastTime != null) {
-            SimpleDateFormat formatter = new SimpleDateFormat(TIMESTAMP_FORMAT);
-            updatedTime = formatter.format(lastTime); // NOSONAR
-        }
-        return updatedTime;
-    }
-
-    public String getLastUpdatedTimeLocalDir(String tableName) {
-        String sql = "select last_update_time_local_dir from POLL_DATA_SYNC_CONFIG where table_name=?";
-        String updatedTime = "";
-        Timestamp lastTime = this.jdbcTemplate.queryForObject(
-                sql,
-                Timestamp.class, tableName);
-        if (lastTime != null) {
-            SimpleDateFormat formatter = new SimpleDateFormat(TIMESTAMP_FORMAT);
-            updatedTime = formatter.format(lastTime); // NOSONAR
-        }
-        return updatedTime;
-    }
-
-    public void updateLastUpdatedTime(String tableName, Timestamp timestamp) {
-        String updateSql = "update POLL_DATA_SYNC_CONFIG set last_update_time =? where table_name=?;";
-        jdbcTemplate.update(updateSql, timestamp, tableName);
-    }
-
-
-    public void updateLastUpdatedTimeAndLog(String tableName, Timestamp timestamp, LogResponseModel logResponseModel) {
-        String updateSql = "update POLL_DATA_SYNC_CONFIG set last_update_time =? where table_name=?;";
-        jdbcTemplate.update(updateSql, timestamp, tableName);
-
-        PollDataLog pollDataLog = new PollDataLog(logResponseModel, tableName);
-        pollDataLogRepository.save(pollDataLog);
-    }
-
-    public void updateLastUpdatedTimeAndLogS3(String tableName, Timestamp timestamp,  LogResponseModel logResponseModel) {
-        String updateSql = "update POLL_DATA_SYNC_CONFIG set last_update_time_s3 =? where table_name=?;";
-        jdbcTemplate.update(updateSql, timestamp, tableName);
-
-        PollDataLog pollDataLog = new PollDataLog(logResponseModel, tableName);
-        pollDataLogRepository.save(pollDataLog);
-    }
-
-    public void updateLastUpdatedTimeAndLogLocalDir(String tableName, Timestamp timestamp, LogResponseModel logResponseModel) {
-        String updateSql = "update POLL_DATA_SYNC_CONFIG set last_update_time_local_dir =? where table_name=?;";
-        jdbcTemplate.update(updateSql, timestamp, tableName);
-
-        PollDataLog pollDataLog = new PollDataLog(logResponseModel, tableName);
-        pollDataLogRepository.save(pollDataLog);
-    }
-
-
-
-    public List<PollDataSyncConfig> getTableListFromConfig() {
-        String sql = "select * from poll_data_sync_config pdsc order by table_order";
-        List<PollDataSyncConfig> tableList = jdbcTemplate.query(
-                sql,
-                new BeanPropertyRowMapper<>(PollDataSyncConfig.class));
-        return tableList;
-    }
-
-    public void deleteTable(String tableName) {
-        try {
-            String deleteSql = "delete FROM " + tableName;
-            jdbcTemplate.execute(deleteSql);
-        } catch (Exception e) {
-            logger.error("RDB:Error in deleting table:{}", e.getMessage());
-        }
-    }
 }
