+++ nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/srte/dao/SrteDataPersistentDAO.java
@@ -36,105 +36,25 @@ import static gov.cdc.nnddatapollservice.share.StringUtil.getStackTraceAsString;
 @Slf4j
 public class SrteDataPersistentDAO {
     private static Logger logger = LoggerFactory.getLogger(SrteDataPersistentDAO.class);
-    private JdbcTemplate jdbcTemplate;
-    @Value("${datasync.sql_error_handle_log}")
-    protected String sqlErrorPath = "";
     @Value("${datasync.data_sync_batch_limit}")
     protected Integer batchSize = 1000;
     private final JdbcTemplateUtil jdbcTemplateUtil;
-    private final HandleError handleError;
-    private final CodeToConditionRepository codeToConditionRepository;
-    private final Gson gsonNorm = new Gson();
-    private final Gson gson = new GsonBuilder()
-            .setFieldNamingPolicy(FieldNamingPolicy.UPPER_CASE_WITH_UNDERSCORES)
-            .create();
     @Autowired
-    public SrteDataPersistentDAO(@Qualifier("rdbJdbcTemplate") JdbcTemplate jdbcTemplate, JdbcTemplateUtil jdbcTemplateUtil, HandleError handleError, CodeToConditionRepository codeToConditionRepository) {
-        this.jdbcTemplate = jdbcTemplate;
+    public SrteDataPersistentDAO(
+            JdbcTemplateUtil jdbcTemplateUtil
+    ) {
         this.jdbcTemplateUtil = jdbcTemplateUtil;
-        this.handleError = handleError;
-        this.codeToConditionRepository = codeToConditionRepository;
     }
 
     protected LogResponseModel handlingSrteTable(PollDataSyncConfig pollConfig, String jsonData, boolean initialLoad) {
         LogResponseModel log = new LogResponseModel();
         log.setLog(LOG_SUCCESS);
         log.setStatus(SUCCESS);
-//        if ("CODE_TO_CONDITION".equalsIgnoreCase(pollConfig.getTableName())) {
-//            Type resultType = new TypeToken<List<CodeToConditionDto>>() {
-//            }.getType();
-//            List<CodeToConditionDto> list = gson.fromJson(jsonData, resultType);
-//            persistingCodeToCondition(list, pollConfig.getTableName());
-//        } else {
-            log = persistingGenericTable(pollConfig, jsonData, initialLoad);
-//        }
+        log = jdbcTemplateUtil.persistingGenericTable(jsonData, pollConfig, initialLoad);
 
         return log;
     }
 
-    protected void persistingCodeToCondition(List<CodeToConditionDto> list, String tableName) {
-        for (CodeToConditionDto data : list) {
-            try {
-                var domainModel = new CodeToCondition(data);
-                codeToConditionRepository.save(domainModel);
-            } catch (Exception e) {
-                logger.error("ERROR occured at record: {}, {}", gsonNorm.toJson(data), e.getMessage()); // NOSONAR
-                handleError.writeRecordToFileTypedObject(gsonNorm, data, tableName + UUID.randomUUID(), sqlErrorPath + "/RDB_MODERN/" + e.getClass().getSimpleName() + "/" + tableName + "/"); // NOSONAR
-            }
-        }
-    }
-
-
-    @SuppressWarnings("java:S3776")
-    protected LogResponseModel persistingGenericTable (PollDataSyncConfig pollConfig, String jsonData, boolean initialLoad) {
-        LogResponseModel log = new LogResponseModel();
-        log.setLog(LOG_SUCCESS);
-        SimpleJdbcInsert jdbcInsert = new SimpleJdbcInsert(jdbcTemplate);
-        jdbcInsert = jdbcInsert.withTableName(pollConfig.getTableName());
-        List<Map<String, Object>> records = PollServiceUtil.jsonToListOfMap(jsonData);
-        if (records != null && !records.isEmpty()) {
-            try {
-                if (records.size() > batchSize) {
-                    int sublistSize = batchSize;
-                    for (int i = 0; i < records.size(); i += sublistSize) {
-                        int end = Math.min(i + sublistSize, records.size());
-                        List<Map<String, Object>> sublist = records.subList(i, end);
-//                        jdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(sublist));
-                        if (initialLoad || pollConfig.getKeyList().isEmpty()) {
-                            sublist.forEach(data -> data.remove("RowNum"));
-                            jdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(sublist));
-
-                        } else {
-                            jdbcTemplateUtil.upsertBatch(pollConfig.getTableName(), sublist, pollConfig.getKeyList());
-                        }
-                    }
-                } else {
-                    jdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(records));
-                }
-            } catch (Exception e) {
-                for (Map<String, Object> res : records) {
-                    try {
-//                        jdbcInsert.execute(new MapSqlParameterSource(res));
-                        if (initialLoad || pollConfig.getKeyList().isEmpty()) {
-                            records.forEach(data -> data.remove("RowNum"));
-                            jdbcInsert.executeBatch(SqlParameterSourceUtils.createBatch(records));
-
-                        } else {
-                            jdbcTemplateUtil.upsertBatch(pollConfig.getTableName(), records, pollConfig.getKeyList());
-                        }
-                    } catch (Exception ei) {
-                        logger.error("ERROR occured at record: {}, {}", gsonNorm.toJson(res), ei.getMessage()); // NOSONAR
-                        handleError.writeRecordToFile(gsonNorm, res, pollConfig.getTableName() + UUID.randomUUID(), sqlErrorPath + "/SRTE/" + ei.getClass().getSimpleName() + "/" + pollConfig.getTableName() + "/");
-                    }
-                }
-            }
-
-        } else {
-            log.setLog("No Data");
-            logger.info("saveSRTEData tableName: {} Records size:0", pollConfig.getTableName());
-        }
-        return log;
-    }
     public LogResponseModel saveSRTEData(PollDataSyncConfig pollConfig, String jsonData, boolean initialLoad) {
         logger.info("saveSRTEData tableName: {}", pollConfig.getTableName());
         LogResponseModel log = new LogResponseModel();
@@ -153,14 +73,5 @@ public class SrteDataPersistentDAO {
         }
 
         return log;
-    }
-
-    public void deleteTable(String tableName) {
-        try {
-            String deleteSql = "delete FROM " + tableName;
-            jdbcTemplate.execute(deleteSql);
-        } catch (Exception e) {
-            logger.error("SRTE:Error in deleting table:{}", e.getMessage());
-        }
     }
 }
